Extracting inspector controls component into a component.
