+++ packages/block-editor/src/style.scss
@@ -29,6 +29,7 @@
 @import "./components/button-block-appender/style.scss";
 @import "./components/colors-gradients/style.scss";
 @import "./components/contrast-checker/style.scss";
+@import "./components/custom-class-name-control/style.scss";
 @import "./components/default-block-appender/style.scss";
 @import "./components/duotone-control/style.scss";
 @import "./components/font-appearance-control/style.scss";
@@ -55,7 +56,6 @@
 @import "./hooks/anchor.scss";
 @import "./hooks/layout.scss";
 @import "./hooks/border.scss";
-@import "./hooks/custom-class-name.scss";
 @import "./hooks/dimensions.scss";
 @import "./hooks/typography.scss";
